requirements.txt
python-dotenv
selenium
webdriver-manager 
wireguard
wireguard-tools